requirements.txt
Flask
PyCryptodome
PyPDF2
qrcode
Pillow
pyzbar